+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.6.2]
+- Fixed descriptions when hovering over types
+- Fixed hover description for `_default` built-in function
+
 ## [0.6.1]
 - Fixed issues with parsing negative number literals